Rewrite swagger parser using Reader monad.
Also, added a level of indirection in responses.
Now there is:
  1) a xxxResponse template and
  2) a xxxResponseBody variant which contains possible responses
This allows pattern matching against the result.

One problem with this is that as DAML data types only support a
single argument, we must add an extra level of indirection to pass
arguments in a single data type. We could use tuples instead, but
it's difficult to track a large number of arguments in this case.